refactor: Restructure files & folders
After team discussion we decided this simpler folder structure would suffice for the forseable
future.
