+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './swatch-bar'
 export * from './product-card';
+export * from './search-box';
+export * from './suggestions';
