+++ src/components/search-box/index.ts
@@ -1,0 +1,3 @@
+export { SearchBox } from './search-box';
+export { useSearchBox } from '../../hooks/search-box.hook';
+export * from './search-box.types';
