+++ src/components/search-box/search-box.scss
