+++ src/components/search-box/search-box.test.stories.tsx
