+++ src/components/search-box/search-box.test.tsx-snapshots/SearchBox-Base-should-render-successfully-1-edge-linux.png
