+++ src/components/search-box/search-box.types.ts
