+++ src/components/suggestions/suggestions.scss
