+++ src/components/suggestions/suggestions.spec.tsx
