+++ src/components/suggestions/suggestions.stories.tsx
@@ -5,11 +5,11 @@ import {
   Configuration,
   ProductSearchOptions,
 } from '@bloomreach/discovery-web-sdk';
-import { LimitlessUIProvider } from '../../core/limitless-ui.provider';
+import { LimitlessUIProvider } from '../../contexts/limitless-ui.provider';
 import { SearchBox } from '../search-box/search-box';
 import { SearchBoxProps } from '../search-box/search-box.types';
 import { Suggestions } from './suggestions';
-import { AutoSuggestContextProvider } from '../context/autosuggest.context';
+import { AutoSuggestContextProvider } from '../../contexts/autosuggest.context';
 
 const meta: Meta<typeof Suggestions> = {
   title: 'SEARCH/Suggestions',
