+++ src/components/suggestions/suggestions.test.tsx-snapshots/Suggestions-Base-should-render-successfully-1-webkit-linux.png
