--- src/core/index.ts
@@ -1,3 +1,0 @@
-export {
-  LimitlessUIProvider
-} from './limitless-ui.provider'
