+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useAutoSuggest } from './autosuggest-api.hook';
+export { useSearch } from './search-api.hook';
