+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './core';
-export * from './search';
+export * from './hooks';
+export * from './contexts';
 export * from './components';
 export * from './utils';
